started distribution-matching approach
Instead of using ID3 with behavior cloning, I'm going to try to copy the behavior of policy gradients with a decision tree. Essentially, the tree policy will output a distribution; at every training step, the target distribution will be slightly adjusted and a new tree will be trained to match the new distribution better.