somehow made cartpole WAY better
I have no idea why this works. All I've done is change the heuristic used to tell which split is best. The greedy action selector still works the way it originally did.

The old heuristic assumed that the mean reward for the greedy action would be representative of the mean reward for that action in general at that place in the tree. The new heuristic simply looks at the total number of times the action was taken under the old policy times the mean reward. Thus, an action won't be favored (by the heuristic) if it was only taken a few times in the old policy, even if its mean reward is higher. However, said actions will be favored when greedily selecting an action. This is kind of a weird mix, and I'm surprised that it works.