track number of features & tree lookup without slice
This was basically a set of improvements around the Sample interface. The old interface had no notion of length, which was annoying in some cases. In other cases, we don't even need a length (e.g. tree lookups), but yet there was no way to do tree lookups on Samples.