batch gradient calcs in Builder
Using batch gradient calculations requires many fewer allocations and interface calls. The performance gained by this varies based on the depth of the tree, since gradient calculations are a one-time computation whereas the rest of the operations have to happen multiple times.

This gives a 1.4x speedup on BenchmarkBuild/ManySamples/Single-8 and a similar speedup (1.6x) on the multithreaded version.