added three new tree algorithms
The BalancedSumAlgorithm was simply an attempt to get the sum algorithm to pass the test defined in TestBuildMSE. The plain old sum algorithm produces a poor solution on that synthetic data.

The StddevAlgorithm was an attempt to match other decision tree implementations. It seems to produce a much more balanced tree than MSEAlgorithm. I have to think harder about MSEAlgorithm to see if there's an inherent reason for it to produce unbalanced trees. (reference: http://www.ke.tu-darmstadt.de/lehre/archiv/ws0809/mldm/dt.pdf, search "Standard deviation")

The SignAlgorithm is the one I'm the most hopeful about. It explicitly tries to produce the best tree when taking the signs of the gradient sums at the leaves. It seems to produce a fairly balanced tree, with two popular leaves encouraging left and right, respectively.