reusable Samples for Judger
Training the value function is a lot slower than training the objective when we use a sample fraction of something like 0.2. Basically, generating the samples from the entire dataset takes more time than building the tree does. Now, we only have to produce a []Sample once and we can reuse it for multiple value steps.