avoid growing slices dynamically
By pre-allocating the slices, we can avoid multiple runtime.growslice calls. According to a profile on pong (with a big batch size), runtime.growslice was taking up 17% of total execution time. Annoyingly, uint8Sample.Feature() was supposedly taking up 36% of total execution time. Perhaps there is a way to optimize for that.