apply forests to data in parallel
On RAM environments, building a tree takes a fraction of a second, but applying 2000 trees to each sample takes ~5 seconds. Since applying the trees was not parallelized before, this meant the training program would be using 100% CPU (rather than n*100% CPU) most of the time (after enough training took place to get a lot of trees).