faster splitTracker for MSEAlgorithm
I found that 9% of the training overhead for Pong-v0 comes from smallVec.Dot() in Reset() and MoveToLeft(). Now the dot products have been moved to Quality(), which is called much less frequently for environments with only a few different possible splits per dimension (0x100 in the case of 8-bit pixels). As the test shows (and as you can prove with algebra), the mean-based formula is equivalent to MSE without needing to compute a rolling square sum.