make objective functions public
This will be necessary when I implement a fine-tuning scheme. With fine-tuning, the user will specify an objective and a forest and get a gradient with respect to the weights in the forest. The objective might come from either a Builder or a PPO object, and the user should be able to specify it using the Objective methods of those objects.